pyproject.toml
[project]
name = "arxiv-papers"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex",email = "alex.kyr.triant@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv (>=2.4.0,<3.0.0)",
    "anthropic (>=0.79.0,<0.80.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
